New version of Copy Item aspect: make group configurable
The previous version of the Copy Item aspect allowed only members of the
Administrators group to deposit a copy of an existing item. This version makes
the group configurable (via copyitem.authorised.groupid in the lconz-aspect
configuration module). This version also overhauls the sitemap to be more
efficient in detecting whether the current page belongs to the Copy Item
aspect.